refactor(daemon): extract command kind logging map
