+++ crates/app/src/tools/mod.rs
@@ -2017,7 +2017,6 @@ mod tests {
             "web.fetch",
             "web.search",
         ]);
-
         assert_eq!(names, expected);
     }
